@@ -8,17 +8,12 @@
 </template>
   
 <script>
-import Bookshelf from './bookshelf.vue'
+import { ref, computed, onMounted, onBeforeUnmount } from 'vue';
+import Bookshelf from './bookshelf.vue';
 
 export default {
     components: {
         'bookshelf': Bookshelf
-    },
-    data() {
-        // Store the current window's width
-        return {
-            windowWidth: window.innerWidth
-        };
     },
     props: {
         // Filtered list of books passed down from App.vue
@@ -27,48 +22,55 @@ export default {
             required: true
         }
     },
-    // Add event listener and update screen width value when the component is mounted
-    mounted() {
-        window.addEventListener('resize', this.updateWidth);
-        this.updateWidth();
-    },
-    // Remove event listener when component is destroyed
-    beforeDestroy() {
-        window.removeEventListener('resize', this.updateWidth);
-    },
-    computed: {
-        // Organise books into shelves based on current size of shelf.
-        shelves() {
-            const shelfSize = this.getShelfSize();
-            const shelves = [];
-            // Split book array to fit on each shelf.
-            for (let i = 0; i < this.filteredBooks.length; i += shelfSize) {
-                shelves.push(this.filteredBooks.slice(i, i + shelfSize));
-            }
-            return shelves;
-        }
-    },
-    methods: {
+    setup(props) {
+        // Store the current window's width
+        const windowWidth = ref(window.innerWidth);
+
+        // Add event listener and update screen width value when the component is mounted
+        onMounted(() => {
+            window.addEventListener('resize', updateWidth);
+            updateWidth();
+        });
+
+        // Remove event listener when component is destroyed
+        onBeforeUnmount(() => {
+            window.removeEventListener('resize', updateWidth);
+        });
+
         // Method to update window width when window is resized
-        updateWidth() {
-            this.windowWidth = window.innerWidth;
-        },
+        const updateWidth = () => {
+            windowWidth.value = window.innerWidth;
+        };
+
         // Method to determine size of shelf based on window width
-        getShelfSize() {
+        const getShelfSize = () => {
             // Return max amount of books to fit on shelf
-            const width = this.windowWidth;
+            const width = windowWidth.value;
             if (width > 1750) return 8;
             if (width > 1500) return 7;
             if (width > 1250) return 6;
             if (width > 1000) return 4;
             if (width > 600) return 3;
             return 2;
-        },
+        };
+
         // Method to select one of four background image for each shelf
-        getShelfBackground(index) {
+        const getShelfBackground = (index) => {
             return (index % 4) + 1;
-        }
-    }
+        };
+
+        // Organise books into shelves based on current size of shelf.
+        const shelves = computed(() => {
+            const shelfSize = getShelfSize();
+            const shelves = [];
+            // Split book array to fit on each shelf.
+            for (let i = 0; i < props.filteredBooks.length; i += shelfSize) {
+                shelves.push(props.filteredBooks.slice(i, i + shelfSize));
+            }
+            return shelves;
+        });
+        return { windowWidth, shelves, updateWidth, getShelfSize, getShelfBackground };
+    },
 }
 </script>
   
